pyproject.toml
[project]
name = "text2sql-langgraph"
version = "0.1.0"
description = "Text2SQL conversion system using LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pymongo>=4.0.0",
    "azure-search-documents>=11.0.0",
    "azure-identity>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "click>=8.0.0",
    "colorlog>=6.7.0",
    "tenacity>=8.2.0",
    "cachetools>=5.3.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "pylint>=3.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "pre-commit>=3.0.0",
]
web = [
    "fastapi>=0.100.0",
    "uvicorn>=0.27.0",
    "streamlit>=1.30.0",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pylint.messages_control]
disable = "C0330, C0326, C0116"
max-line-length = 88

[tool.pylint.format]
max-module-lines = 1000

[tool.pylint.basic]
good-names = "i,j,k,ex,_,df,e"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
pythonpath = [
    "src",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]